d710: Fix magnetometer axis configuration
Although the i9100 and the d710 share the same sensor
circuit design, the circuit board layout is different
and so is the orientation of sensors relative to that
of the device. This means that sensor-to-device axis
and direction mappings differ between the two devices.

Accelerometer and gyro fix here:
http://review.cyanogenmod.org/#/c/86259/

Test details here:
http://forum.xda-developers.com/epic-4g-touch/orig-development/rom-cyanogenmod-11-official-nightly-t2852074/post58297106

Thanks to Kjricker on XDA who ran measurements on the
device and tested the fix.

Change-Id: I4b69c3b7cd8911f4cf182b709bc2d441f4efac46